consumer/requirements.txt
kafka-python>=2.0.2,<3.0.0
mysql-connector-python>=8.0.0,<9.0.0
